backend/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
pandas
numpy
torch
ultralytics
joblib
scikit-learn
